# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Documentation Articles",
        "description": "Codelist Technical Test API Documentation Muhammad Arif",
        "contact": {
            "email": "areef.empat.empat@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/articles": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Menampilkan Semua List Artikel",
                "description": "List Artikel",
                "operationId": "11aa4cc6561a7b94c56d1201d62513ee",
                "responses": {
                    "default": {
                        "description": "List Articles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Membuat Artikel Baru",
                "description": "Endpoint untuk membuat artikel baru",
                "operationId": "e328bed23caf889ad982ad3f1180bccb",
                "requestBody": {
                    "description": "Data artikel yang akan dibuat",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "content",
                                    "id_categories",
                                    "media"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "id_categories": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Artikel baru berhasil dibuat"
                    }
                }
            }
        },
        "/api/articles/{id}": {
            "get": {
                "tags": [
                    "Articles"
                ],
                "summary": "Menampilkan 1 Artikel Berdasarkan id",
                "description": "Details Articles",
                "operationId": "b3e529eab04efb5d305b07e2e58d9869",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id artikel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Details Articles"
                    }
                }
            },
            "post": {
                "tags": [
                    "Articles"
                ],
                "summary": "Memperbaharui 1 Artikel Baru berdasarkan id",
                "description": "Endpoint untuk Memperbaharui artikel",
                "operationId": "5450592915242f040336c5fc74897459",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id artikel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Data artikel yang akan dibuat",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "title",
                                    "author",
                                    "content",
                                    "id_categories",
                                    "media"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "author": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "id_categories": {
                                        "type": "string"
                                    },
                                    "media": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Artikel baru berhasil dibuat"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Articles"
                ],
                "summary": "Menghapus 1 data Artikel Berdasarkan id",
                "description": "Hapus Artikel",
                "operationId": "97f34592611a1dd53a086865e29706c9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id artikel",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Hapus Artikel"
                    }
                }
            }
        },
        "/api/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Menampilkan Semua List Kategori",
                "description": "List Categories",
                "operationId": "3f5817a34833d0a1f4af4548dd3aeaba",
                "responses": {
                    "default": {
                        "description": "List Categories"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Membuat Kategori Baru",
                "description": "Membuat Kategori Baru",
                "operationId": "71fcad552bb0eaba9fb191fd8d8dcab0",
                "requestBody": {
                    "description": "Menambahkan Kategori Dan Deskripsi Dari Kategori",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "category",
                                    "description"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Kategori Baru Berhasil Dibuat"
                    }
                }
            }
        },
        "/api/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Menampilkan 1 Kategori Berdasarkan id",
                "description": "Details Categories",
                "operationId": "c68e76d323c008827a9e47398b1583de",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id kategori",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Details Categories"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Memperbaharui 1 data Kategori Berdasarkan id",
                "description": "Merubah Kategori",
                "operationId": "0e686b2748211cc688333ed705dc111f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id kategori",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Merubah Kategori Dan Deskripsi",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "category",
                                    "description"
                                ],
                                "properties": {
                                    "category": {
                                        "type": "string"
                                    },
                                    "description": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "default": {
                        "description": "Kategori Berhasil Diubah"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Menghapus 1 data Kategori Berdasarkan id",
                "description": "Hapus Categories",
                "operationId": "4c12e22a7f8c617bd83bfa1fdc05428d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id kategori",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        }
                    }
                ],
                "responses": {
                    "default": {
                        "description": "Hapus Categories"
                    }
                }
            }
        }
    }
}